+++ tests/test_requesters.py
@@ -14,20 +14,8 @@ class FutpediaRequesterTests(unittest.TestCase):
 	"""Set of unit tests to validate an instance of a FutpediaRequester and
 	its methods.
 
-	Tests: test_retry_limit, test_fetch
+	Tests: test_fetch
 	"""
-	def test_retry_limit(self):
-		"""Steps:
-		1 - Instantiates a FutpediaRequester
-		2 - Tests the getter
-		3 - Tests the setter
-		"""
-		with FutpediaRequester(retry_limit=3) as requester:
-			self.assertEqual(requester.retry_limit, 3)
-
-			requester.retry_limit = 7
-			self.assertEqual(requester.retry_limit, 7)
-
 	def test_fetch(self):
 		"""Steps:
 		1 - Instantiates a FutpediaRequester
@@ -38,8 +26,7 @@ class FutpediaRequesterTests(unittest.TestCase):
 			res = requester.fetch('/')
 			self.assertIsInstance(res, bytes)
 
-			requester.retry_limit = 1
-
+		with FutpediaRequester(retry_limit=1) as requester:
 			with self.assertRaises(ScrapediaRequestError):
 				requester.fetch('/unknown')